Updated By Sharat -Lorem Ipsum
Lorem ipsum is the nonsense filler text that typically demonstrates the font and style of a text in a document or visual demonstration. Originally from Latin, lorem ipsum has no intelligible meaning, but is simply a display of letters and characteristics to be viewed as a sample with given graphical elements in a file. Lipsum (portmanteau of lorem and ipsum) generators are commonly used to form generic text in a file. The “gibberish” is adequately like normal text to demonstrate a font, but does not distract the reader with its content. It has been used as placeholder text since the 16th century.